Fix software rows saved without a software title errors
Resolves #51913

Revert the concurrent title-ingestion changes from #49894, keeping only
the batching of host_software_installed_paths deletes (#49805).

deduplicated by a per-instance cache with no cross-instance
invalidation. After CleanupSoftwareTitles on one Fleet instance
garbage-collected a title, other instances skipped re-inserting it and
wrote software rows with a NULL title_id — invisible to every query that
joins software to software_titles, and never repaired. Titles are now
inserted inside the ingestion transaction again, so a title cannot be
deleted between insert and resolution.
